Split the native crate into cohesive modules and enforce the size rule for Rust
Pure refactor of the gateway data plane: server.rs, waterfall.rs, and
dialects.rs move verbatim into cohesive modules, each under the repo's
999-line limit. Behavior is locked by the unchanged crate unit tests,
python parity and golden fixtures, and gateway integration suites.

Add .rs to HAND_AUTHORED_SUFFIXES so oversized Rust files fail the
layout gate, and bump the crate to 0.1.13 in lockstep (Cargo.toml,
crate pyproject, flagship floor, Cargo.lock, uv.lock).
